+++ axelor-account/src/main/java/com/axelor/apps/account/service/invoice/generator/InvoiceLineGenerator.java
@@ -24,6 +24,7 @@ import com.axelor.apps.account.db.InvoiceLine;
 import com.axelor.apps.account.db.Tax;
 import com.axelor.apps.account.db.TaxEquiv;
 import com.axelor.apps.account.db.TaxLine;
+import com.axelor.apps.account.db.repo.InvoiceLineRepository;
 import com.axelor.apps.account.exception.IExceptionMessage;
 import com.axelor.apps.account.service.AccountManagementAccountService;
 import com.axelor.apps.account.service.app.AppAccountService;
@@ -87,10 +88,11 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
 	protected int discountTypeSelect;
 	protected BigDecimal exTaxTotal;
 	protected BigDecimal inTaxTotal;
-	protected boolean isTitleLine;
+	protected Integer typeSelect = 0;
+	protected boolean isSubLine;
 
 	public static final int DEFAULT_SEQUENCE = 0;
-
+	
 	protected InvoiceLineGenerator() { }
 
 	protected InvoiceLineGenerator( Invoice invoice ) {
@@ -104,7 +106,7 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
     }
 
 	protected InvoiceLineGenerator( Invoice invoice, Product product, String productName, String description, BigDecimal qty,
-			Unit unit, int sequence, boolean isTaxInvoice) {
+			Unit unit, int sequence, boolean isTaxInvoice, boolean isSubLine) {
 		
 		this(invoice);
 		
@@ -117,13 +119,14 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
         this.isTaxInvoice = isTaxInvoice;
         this.today = Beans.get(AppAccountService.class).getTodayDate();
         this.currencyService = new CurrencyService(this.today);
+        this.isSubLine = isSubLine;
 	}
 	
 	protected InvoiceLineGenerator( Invoice invoice, Product product, String productName, BigDecimal price, BigDecimal priceDiscounted, String description, BigDecimal qty,
 			Unit unit, TaxLine taxLine, int sequence, BigDecimal discountAmount, int discountTypeSelect, BigDecimal exTaxTotal,
-			BigDecimal inTaxTotal, boolean isTaxInvoice) {
+			BigDecimal inTaxTotal, boolean isTaxInvoice, boolean isSubLine) {
 
-        this(invoice, product, productName, description, qty, unit, sequence, isTaxInvoice);
+        this(invoice, product, productName, description, qty, unit, sequence, isTaxInvoice, isSubLine);
 		
         this.price = price;
         this.priceDiscounted = priceDiscounted;
@@ -172,7 +175,7 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
 
 		invoiceLine.setDescription(description);
 		invoiceLine.setPrice(price);
-
+		invoiceLine.setIsSubLine(isSubLine);
 		invoiceLine.setPriceDiscounted(priceDiscounted);
 		invoiceLine.setQty(qty);
 		invoiceLine.setUnit(unit);
@@ -209,7 +212,7 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
 		invoiceLine.setDiscountTypeSelect(discountTypeSelect);
 		invoiceLine.setDiscountAmount(discountAmount);
 		
-		invoiceLine.setIsTitleLine(isTitleLine);
+		invoiceLine.setTypeSelect(typeSelect);
 
 		return invoiceLine;
 
@@ -229,7 +232,7 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
 
 	public void computeTotal()  {
 		
-		if(isTitleLine)  {  return;  }
+		if(typeSelect == InvoiceLineRepository.TYPE_TITLE)  {  return;  }
 		
 		BigDecimal taxRate = BigDecimal.ZERO;
 		if(taxLine != null)  {  taxRate = taxLine.getValue();  }
@@ -246,7 +249,7 @@ public abstract class InvoiceLineGenerator extends InvoiceLineManagement {
 	
 	public void computeCompanyTotal(InvoiceLine invoiceLine) throws AxelorException  {
 		
-		if(isTitleLine)  {  return;  }
+		if(typeSelect == InvoiceLineRepository.TYPE_TITLE)  {  return;  }
 		
 		Company company = invoice.getCompany();