+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/expense/ExpenseServiceImpl.java
@@ -515,7 +515,7 @@ public class ExpenseServiceImpl implements ExpenseService {
 		if (atiChoice == 1 || atiChoice == 3) {
 			invoiceLineGenerator = new InvoiceLineGenerator(invoice, product, product.getName(), expenseLine.getUntaxedAmount(),
 					expenseLine.getUntaxedAmount(), expenseLine.getComments(), BigDecimal.ONE, product.getUnit(), null, priority, BigDecimal.ZERO, IPriceListLine.AMOUNT_TYPE_NONE,
-					expenseLine.getUntaxedAmount(), expenseLine.getTotalAmount(), false) {
+					expenseLine.getUntaxedAmount(), expenseLine.getTotalAmount(), false, false) {
 
 				@Override
 				public List<InvoiceLine> creates() throws AxelorException {
@@ -531,7 +531,7 @@ public class ExpenseServiceImpl implements ExpenseService {
 		} else {
 			invoiceLineGenerator = new InvoiceLineGenerator(invoice, product, product.getName(), expenseLine.getTotalAmount(),
 					expenseLine.getTotalAmount(), expenseLine.getComments(), BigDecimal.ONE, product.getUnit(), null, priority, BigDecimal.ZERO, IPriceListLine.AMOUNT_TYPE_NONE,
-					expenseLine.getUntaxedAmount(), expenseLine.getTotalAmount(), false) {
+					expenseLine.getUntaxedAmount(), expenseLine.getTotalAmount(), false, false) {
 
 				@Override
 				public List<InvoiceLine> creates() throws AxelorException {